seperated business and businesses redux stores
