+++ client/store/business/getBusiness.js
@@ -1,23 +1,16 @@
 import axios from 'axios'
 
-const GET_BUSINESS = 'GET_BUSINESS'
+const GET_BUSINESSES = 'GET_BUSINESSES'
 
 // action creator
-const getBusiness = business => ({type: GET_BUSINESS, business})
+const getBusinesses = businesses => ({type: GET_BUSINESSES, businesses})
 
 // thunk
-export const fetchBusiness = () => async dispatch => {
-  try {
-    const {data: businessData} = await axios.get('/api/business/portal')
-    dispatch(getBusiness(businessData))
-  } catch (error) {
-    console.log('error in fetchBusiness', error)
-  }
-}
+
 export const getAllBussiness = () => async dispatch => {
   try {
     const {data: businessData} = await axios.get('/api/business')
-    dispatch(getBusiness(businessData))
+    dispatch(getBusinesses(businessData))
   } catch (error) {
     console.log('error in fetchBusiness', error)
   }
@@ -25,8 +18,8 @@ export const getAllBussiness = () => async dispatch => {
 
 export const searchBusiness = query => async dispatch => {
   try {
-    const {data: business} = await axios.get(`/api/search?search=${query}`)
-    dispatch(getBusiness(business))
+    const {data: businesses} = await axios.get(`/api/search?search=${query}`)
+    dispatch(getBusinesses(businesses))
   } catch (error) {
     console.log('error in fetchProducts', error)
   }
@@ -36,20 +29,20 @@ export const getCategory = query => async dispatch => {
     const {data: category} = await axios.get(
       `/api/business/filter/category?category=${query}`
     )
-    dispatch(getBusiness(category))
+    dispatch(getBusinesses(category))
   } catch (error) {
     console.log('error in category product thunk', error)
   }
 }
 
 // reducer
-const business = (state = [], action) => {
+const businesses = (state = [], action) => {
   switch (action.type) {
-    case GET_BUSINESS:
-      return action.business
+    case GET_BUSINESSES:
+      return action.businesses
     default:
       return state
   }
 }
 
-export default business
+export default businesses
