+++ client/store/index.js
@@ -4,16 +4,18 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 import customer from './customer'
-import business from './business/getBusiness'
+import businesses from './business/getBusiness'
 import businessCreate from './business/createBusiness'
 import subscription from './business/subscription'
+import business from './business/singleBusiness'
 
 const reducer = combineReducers({
   user,
   customer,
   business,
   businessCreate,
-  subscription
+  subscription,
+  businesses
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
